don't log empty JavaScript exception (#13480)
In case we log a JavaScript exception to the logfile, unrecoverable
exceptions without a message currently log "JavaScript exception:",
which looks like a bug.
For example, this happens if the arangosh is ended by a script with something similar to `process.exit(1)`.
In this case, the exception that this thrown internally is
unrecoverable, and does not have any message attached. So we can as well
skip logging it.